refactor(engine): extract spill/buffer.rs (SPL-13 #2411)
